requirements.txt
# Web Framework
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20

# Database & Vector Store
pymongo==4.13.2
motor==3.7.1
qdrant-client==1.15.0

# AI & Language Models
openai==1.97.0
langchain==0.3.26
langchain-core==0.3.70
langchain-openai==0.3.28
litellm==1.74.7

# MCP Protocol
mcp==1.12.0

# Document Processing
PyMuPDF==1.26.3
pymupdf4llm==0.0.27
python-docx==1.2.0
pdf2image==1.16.3

# OCR & Image Processing
pytesseract==0.3.13
Pillow==10.4.0
opencv-python-headless>=4.8.0

# Text Processing & NLP
tiktoken==0.9.0
semchunk==3.2.2
regex==2024.11.6
lingua-language-detector==2.1.1
spacy==3.8.7
nltk==3.9.1

numpy>=1.24.0,<2.3.0
pandas==2.3.1
scikit-learn>=1.3.0
scipy>=1.11.0

# Configuration & Environment
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1

# Utilities
tenacity==9.1.2
python-dateutil==2.9.0.post0

#Azure
azure-ai-projects~=1.0.0b11
azure-identity~=1.20.0
ansible-core~=2.17.0
